Add Prediction logic with `predictor` module
This commit adds a module `predictor.py`, which contains the function `predict()` returning the predicted class with the probability of prediction. It contains a `load_model()` function to load the models and the class names. Plus two functions `get_cropped_image_if_2_eyes` and `w2d` for image processing and creating a wavelet form of the image. The `predict()` function vertically stacks these images for the prediction.