Refactor Prediction logic
This commit refactors the prediction logic in the `predict()` function by only returning the predicted class and the probability instead of all the encoded classes. Also, an error check has been handled if no faces were detected. In that case, the function returns the predicted class 'No face detected' with 0% probability.