Integrate Streamlit Web Application to `predictor` module
This commit integrates the Streamlit web application to the `predictor` module. The `load_model()` function is called on page load even before setting the page configurations and drawing the UI and the `predict()` function is called whenever and image is uploaded. Prediction results are shown if a face can be detected otherwise the message 'No face detected' is shown.

The UI of the web application has also been built with the possible celebrities that can be identified shown in the page itself.